requirements.txt
streamlit
aiohttp
python-dotenv